Move shared helper functions to Export DOS opportunities
This is the only other script that was using them.
