--- dmscripts/models/process_rules.py
@@ -1,10 +1,0 @@
-from datetime import datetime
-from dmutils.formats import DATE_FORMAT, DATETIME_FORMAT
-
-
-def format_datetime_string_as_date(dt):
-    return datetime.strptime(dt, DATETIME_FORMAT).strftime(DATE_FORMAT) if dt else None
-
-
-def remove_username_from_email_address(ea):
-    return '{}'.format(ea.split('@').pop()) if ea else None
